refactored: deleted and reorganized code
