+++ src/components/atom/Icons/index.jsx
@@ -1,2 +1,1 @@
 export * from "./fontawesomeIcon";
-export * from "./logoIcon";
