--- src/components/atom/Icons/logoIcon.jsx
@@ -1,6 +1,0 @@
-// styling
-import styles from "../../../styles/atom/icons.module.css";
-
-export const logoIcon = (icon) => (
-  <img src={icon} alt={icon} className={styles.logo} />
-);
