+++ src/components/atom/elements/index.jsx
@@ -1,9 +1,7 @@
 export * from "./buttonClick";
-export * from "./displayErrors";
 export * from "./h4Heading";
 export * from "./heading";
 export * from "./image";
 export * from "./subheading";
-export * from "./submitButton";
 export * from "./thumbnail";
 export * from "./title";
