--- src/components/atom/elements/submitButton.jsx
@@ -1,7 +1,0 @@
-import Button from "react-bootstrap/Button";
-
-export const submitButton = (variant) => (
-  <Button variant={variant} type="submit">
-    Submit
-  </Button>
-);
