+++ src/components/pages/LandingPage.jsx
@@ -1,7 +1,7 @@
 // components
-import PageTitleMolecule from "../molecule/PageTitleMolecule";
 import CarouselOrganism from "../organism/CarouselOrganism";
 import TestimonialsOrganism from "../organism/TestimonialsOrganism";
+import PageTitleMolecule from "../molecule/PageTitleMolecule";
 
 const LandingPage = () => {
   return (
