+++ src/components/pages/dashboard/AdminPropertyPage.jsx
@@ -1,6 +1,8 @@
+// components
 import AdminButtonsMolecule from "../../molecule/AdminButtonsMolecule";
 import { heading } from "../../atom/elements";
 
+// styling
 import styles from "../../../styles/pages/admin.module.css";
 
 const AdminPropertyPage = () => {
