+++ src/components/templates/LayoutTemplate.jsx
@@ -1,11 +1,15 @@
+// 3rd party libraries
+import { Alert } from "react-bootstrap";
+
 // components
 import HeaderOrganism from "../organism/HeaderOrganism";
 import FooterOrganism from "../organism/FooterOrganism";
 
+// custom hooks
+import useAppContext from "../../hooks/useAppContext";
+
 // styling
 import styles from "../../styles/templates/Layout.module.css";
-import { Alert } from "react-bootstrap";
-import useAppContext from "../../hooks/useAppContext";
 
 /**
  * This template will be used to render Header, Footer and all content in between
